+++ server/camcops_server/alembic/versions/0020_pbq.py
@@ -40,7 +40,6 @@ Creation date: 2019-04-01 23:52:57.174006
 
 from alembic import op
 import sqlalchemy as sa
-import cardinal_pythonlib.sqlalchemy.list_types
 import camcops_server.cc_modules.cc_sqla_coltypes