+++ server/camcops_server/alembic/versions/0021_hidden_specialnote.py
@@ -1,0 +1,67 @@
+#!/usr/bin/env python
+
+"""
+camcops_server/alembic/versions/0021_hidden_specialnote.py
+
+===============================================================================
+
+    Copyright (C) 2012-2019 Rudolf Cardinal (rudolf@pobox.com).
+
+    This file is part of CamCOPS.
+
+    CamCOPS is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    CamCOPS is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with CamCOPS. If not, see <http://www.gnu.org/licenses/>.
+
+===============================================================================
+
+DATABASE REVISION SCRIPT
+
+hidden_specialnote
+
+Revision ID: 0021
+Revises: 0020
+Creation date: 2019-04-04 11:15:17.797999
+
+"""
+
+# =============================================================================
+# Imports
+# =============================================================================
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# =============================================================================
+# Revision identifiers, used by Alembic.
+# =============================================================================
+
+revision = '0021'
+down_revision = '0020'
+branch_labels = None
+depends_on = None
+
+
+# =============================================================================
+# The upgrade/downgrade steps
+# =============================================================================
+
+# noinspection PyPep8,PyTypeChecker
+def upgrade():
+    with op.batch_alter_table('_special_notes', schema=None) as batch_op:
+        batch_op.add_column(sa.Column('hidden', sa.Boolean(), nullable=False, comment='Manually hidden (effectively: deleted)'))
+
+
+def downgrade():
+    with op.batch_alter_table('_special_notes', schema=None) as batch_op:
+        batch_op.drop_column('hidden')